requirements.txt
backoff==1.4.3
boto3==1.4.7
botocore==1.7.29
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
docker==2.5.1
docker-pycreds==0.2.1
docutils==0.14
Flask==0.12.2
Flask-LogConfig==0.4.2
Flask-Script==2.0.6
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
logconfig==0.4.0
logutils==0.3.5
MarkupSafe==1.0
pymongo==3.5.1
python-dateutil==2.6.1
python-dotenv==0.7.1
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
